smartooth-ai-ML/requirements.txt
flask==2.3.2
flask-cors==3.0.10
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.2.0
gunicorn==20.1.0
prometheus-client==0.17.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
faker==18.11.2
